completed

+++ UseGithubAPI/View/ViewController.swift
@@ -10,44 +10,45 @@ import UIKit
 import Alamofire
 import SwiftyJSON
 import SDWebImage
+import CoreData
 
 class ViewController: UIViewController, UITableViewDataSource, UITableViewDelegate, UISearchBarDelegate {
-
+    
     @IBOutlet weak var forkSort: UIButton!
     @IBOutlet weak var starSort: UIButton!
     @IBOutlet weak var searchBar: UISearchBar!
     @IBOutlet weak var tableInfo: UITableView!
     var refreshControl = UIRefreshControl()
     let getData = GetData(baseUrl: "https://api.github.com/search/repositories")
-    var arrFilter = [GitRepo]()
-    var searching = false
     var page: Int = 1
     var limit: Int = 50
     var desc = false
     var mode: Int = 2
     var sortType: String = ""
     
+    var arrRepoCacheDefault = [GitRepo]()
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         tableInfo.delegate = self
         tableInfo.dataSource = self
         searchBar.delegate = self
-        getData.fetchData(endpoint: "?q=language:&per_page=50&page=\(page)", table: tableInfo)
+        if !InternetCheck.isConnectedToInternet() {
+            print("no internet 00")
+            arrRepoCacheDefault = Contains.cachePublic.get()
+            Contains.arrRepo = arrRepoCacheDefault
+        } else {
+            getData.fetchData(endpoint: "?q=language:&per_page=50&page=\(page)", table: tableInfo)
+        }
         pullToRefresh()
         starSort.addTarget(self, action: #selector(star), for: .touchUpInside)
         forkSort.addTarget(self, action: #selector(fork), for: .touchUpInside)
-        
         tableInfo.tableFooterView = UIView(frame: .zero)
-        var index = 0
-        while index < limit {
-            index = index + 1
-        }
+        
     }
-    
     func pullToRefresh() {
         refreshControl.attributedTitle = NSAttributedString(string: "Refresh")
         refreshControl.addTarget(self, action: #selector(refreshData(_:)), for: .valueChanged)
-        
         if #available(iOS 10.0, *) {
             self.tableInfo.refreshControl = refreshControl
         } else {
@@ -55,59 +56,18 @@ class ViewController: UIViewController, UITableViewDataSource, UITableViewDelega
         }
     }
     @objc private func refreshData(_: Any) {
+        Contains.arrRepo.removeAll()
         tableInfo.dataSource = self
+        tableInfo.delegate = self
         getData.fetchData(endpoint: "?q=language:&per_page=50&page=\(page)", table: tableInfo)
         refreshControl.endRefreshing()
     }
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        if searching {
-            return arrFilter.count
-        } else {
         return Contains.arrRepo.count
-        }
     }
-    func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
-        let lastSectionIndex = tableView.numberOfSections - 1
-        let lastRowIndex = tableView.numberOfRows(inSection: lastSectionIndex) - 1
-        if indexPath.section ==  lastSectionIndex && indexPath.row == lastRowIndex {
-            let spinner = UIActivityIndicatorView(style: .gray)
-            spinner.startAnimating()
-            spinner.frame = CGRect(x: CGFloat(0), y: CGFloat(0), width: tableView.bounds.width, height: CGFloat(44))
-
-            self.tableInfo.tableFooterView = spinner
-            self.tableInfo.tableFooterView?.isHidden = false
-        }
-        if indexPath.row == Contains.arrRepo.count - 1 {
-            if Contains.arrRepo.count < 500 {
-                var index = Contains.arrRepo.count
-                limit = index + 50
-                page += 1
-                while index < limit {
-                    index += 1
-                }
-            }
-             self.perform(#selector(loadMore), with: nil, afterDelay: 0.01)
-        }
-       
-    }
-    @objc func loadMore() {
-        getData.fetchData(endpoint: "?q=language:&per_page=50&page=\(page)", table: tableInfo)
-        debugPrint("loaddata \(page) \(limit)")
-    }
-
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell:TableViewCell = tableInfo.dequeueReusableCell(withIdentifier: "Cell") as! TableViewCell
-        if searching {
-            let repo = arrFilter[indexPath.row]
-            cell.lbName.text = repo.login
-            cell.lbRepo.text = repo.name
-            cell.imgAvatar.sd_setImage(with: URL(string: repo.avatar_url))
-            cell.lbStar.text = repo.star
-            cell.lbWatch.text = repo.watch
-            cell.lbFork.text = repo.fork
-            cell.lbIssue.text = repo.issue
-            cell.lbCommit.text =  "Commit at: \(repo.commit)"
-        } else {
+        if !Contains.arrRepo.isEmpty{
             let repo = Contains.arrRepo[indexPath.row]
             cell.lbName.text = repo.login
             cell.lbRepo.text = repo.name
@@ -116,23 +76,54 @@ class ViewController: UIViewController, UITableViewDataSource, UITableViewDelega
             cell.lbWatch.text = repo.watch
             cell.lbFork.text = repo.fork
             cell.lbIssue.text = repo.issue
-            cell.lbCommit.text =  "Commit at: \(repo.commit)"
-        }
+            cell.lbCommit.text =  "Commit at: \(repo.commit)"}
+        //        }
         return cell
     }
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 180.5
+        return 160
     }
     func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
-//        arrFilter = Contains.arrRepo.filter({$0.login.lowercased().prefix(searchText.count) == searchText.lowercased()})
-        arrFilter = Contains.arrRepo.filter({$0.name.lowercased().prefix(searchText.count) == searchText.lowercased()})
-        searching = true
+        //        arrFilter = Contains.arrRepo.filter({$0.login.lowercased().prefix(searchText.count) == searchText.lowercased()})
+        Contains.arrRepo = Contains.arrRepo.filter({$0.name.lowercased().prefix(searchText.count) == searchText.lowercased()})
+        mode = 3
+        if searchText.isEmpty {
+            Contains.arrRepo = arrRepoCacheDefault
+        }
         tableInfo.reloadData()
     }
-    func searchBarCancelButtonClicked(_ searchBar: UISearchBar) {
-        searching = false
-        searchBar.text = ""
-        tableInfo.reloadData()
+    func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
+        let lastSectionIndex = tableView.numberOfSections - 1
+        let lastRowIndex = tableView.numberOfRows(inSection: lastSectionIndex) - 1
+        if indexPath.section ==  lastSectionIndex && indexPath.row == lastRowIndex {
+            let spinner = UIActivityIndicatorView(style: .gray)
+            spinner.startAnimating()
+            spinner.frame = CGRect(x: CGFloat(0), y: CGFloat(0), width: tableView.bounds.width, height: CGFloat(44))
+            self.tableInfo.tableFooterView = spinner
+            self.tableInfo.tableFooterView?.isHidden = false
+        }
+        let index = Contains.arrRepo.count
+        if indexPath.row == index - 1 && Contains.loadMore {
+            print("indexpath:\(indexPath.row), \(index)")
+            if index < 950 {
+                limit = index + 50
+                page += 1
+                Contains.loadMore = false
+                self.perform(#selector(loadMore), with: nil, afterDelay: 0.01)
+            }
+        }
+    }
+    
+    @objc func loadMore() {
+        if mode == 2 {
+            getData.fetchData(endpoint: "?q=language:&per_page=50&page=\(page)", table: tableInfo)
+        } else if mode == 0 {
+            getData.fetchData(endpoint: "?q=language:&per_page=\(limit)&page=\(page)&sort=stars&order=\(sortType)", table: tableInfo)
+        } else if mode == 1 {
+            getData.fetchData(endpoint: "?q=language:&per_page=\(limit)&page=\(page)&sort=forks&order=\(sortType)", table: tableInfo)
+        } else if mode == 3 {
+            getData.fetchData(endpoint: "?q=language:&per_page=50&page=\(page)", table: tableInfo)
+        }
     }
     @objc func star() {
         //&sort=stars
@@ -142,14 +133,28 @@ class ViewController: UIViewController, UITableViewDataSource, UITableViewDelega
             mode = 0
             desc = true
         }
-        if desc {
-            sortType = "desc"
+        if InternetCheck.isConnectedToInternet() {
+            page = 1
+            if desc {
+                sortType = "desc"
+            } else {
+                sortType = "asc"
+            }
+            tableInfo.dataSource = self
+            Contains.arrRepo.removeAll()
+            tableInfo.reloadData()
+            getData.fetchData(endpoint: "?q=language:&per_page=\(limit)&page=\(page)&sort=stars&order=\(sortType)", table: tableInfo)
         } else {
-            sortType = "asc"
+            starCache()
         }
-        tableInfo.dataSource = self
-        Contains.arrRepo.removeAll()
-        getData.fetchData(endpoint: "?q=language:&per_page=50&page=\(page)&sort=stars&order=\(sortType)", table: tableInfo)
+    }
+    func starCache() {
+        if desc {
+            Contains.arrRepo = Contains.arrRepo.sorted(by: { (Int($0.star) ?? 0) > (Int($1.star) ?? 0) })
+        } else {
+            Contains.arrRepo = Contains.arrRepo.sorted(by: { (Int($0.star) ?? 0) < (Int($1.star) ?? 0) })
+        }
+        tableInfo.reloadData()
     }
     @objc func fork() {
         //&sort=forks
@@ -159,20 +164,34 @@ class ViewController: UIViewController, UITableViewDataSource, UITableViewDelega
             mode = 1
             desc = true
         }
-        if desc {
-            sortType = "desc"
+        if InternetCheck.isConnectedToInternet(){
+            page = 1
+            if desc {
+                sortType = "desc"
+            } else {
+                sortType = "asc"
+            }
+            tableInfo.dataSource = self
+            Contains.arrRepo.removeAll()
+            tableInfo.reloadData()
+            getData.fetchData(endpoint: "?q=language:&per_page=\(limit)&page=\(page)&sort=forks&order=\(sortType)", table: tableInfo)
         } else {
-            sortType = "asc"
+            forkCache()
         }
-        tableInfo.dataSource = self
-          Contains.arrRepo.removeAll()
-        getData.fetchData(endpoint: "?q=language:&per_page=50&page=\(page)&sort=forks&order=\(sortType)", table: tableInfo)
+    }
+    func forkCache () {
+        if desc {
+            Contains.arrRepo = Contains.arrRepo.sorted(by: { (Int($0.fork) ?? 0) > (Int($1.fork) ?? 0) })
+        } else {
+            Contains.arrRepo = Contains.arrRepo.sorted(by: { (Int($0.fork) ?? 0) < (Int($1.fork) ?? 0) })
+        }
+        tableInfo.reloadData()
     }
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let web = storyboard?.instantiateViewController(withIdentifier: "ViewController2") as? ViewController2
         web?.url = Contains.arrRepo[indexPath.row].url
         self.navigationController?.pushViewController(web!, animated: true)
-
+        
     }
 }
